src/requirements.txt
# Core Dependencies
langchain>=0.1.0
chromadb>=0.4.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
tenacity>=8.2.3

# Vector Store
chromadb>=0.4.0
faiss-cpu>=1.7.4  # または faiss-gpu (GPUサポート用)

# PDF Processing
pdfminer.six>=20221105
pypdf>=3.17.1

# UI and Progress
rich>=13.3.1
tqdm>=4.66.1

# Utils
tenacity>=8.2.3

# Development Dependencies
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
types-requests>=2.31.0
types-PyPDF2>=3.0.0
types-python-dateutil>=2.8.19